+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/DukConstants.java
@@ -10,9 +10,9 @@ import org.firstinspires.ftc.robotcore.external.navigation.Orientation;
 public class DukConstants {
     public static final class INPUT {
         public static final boolean MAXIMIZE_MAGNITUDE = true;
-        public static final float MANUAL_DRIVE_CONTROL_MULTIPLIER = 0.5f;
-        public static final float MANUAL_TURN_CONTROL_MULTIPLIER = 0.05f;
-        public static final float JOYSTICK_TURN_THRESHOLD = 0.5f;
+        public static final double MANUAL_DRIVE_CONTROL_MULTIPLIER = 0.5f;
+        public static final double MANUAL_TURN_CONTROL_MULTIPLIER = 0.05f;
+        public static final double JOYSTICK_TURN_THRESHOLD = 0.5f;
     }
 
     public static final class ORIENTATION {
@@ -22,7 +22,7 @@ public class DukConstants {
         public static final Orientation NULL_ORIENTATION = new Orientation(
                 AXES_REFERENCE, AXES_ORDER, ANGLE_UNIT,
                 0, 0, 0, 0);
-        public static final float IMU_CHECK_INTERVAL = 1f;
+        public static final double IMU_CHECK_INTERVAL = 1f;
     }
 
     public static final class DEBUG {
@@ -45,16 +45,16 @@ public class DukConstants {
         //Intersection of parallel center axis with perpendicular axis
         public static final Vector ODOMETER_CENTER = new Vector(0, -1000);
         public static final double ROBOT_SIZE_IN = 15;
-        public static final Vector LEFT_WHEEL_PAIR_PROFILE = new Vector((float)Math.sqrt(2),(float)Math.sqrt(2));
-        public static final Vector RIGHT_WHEEL_PAIR_PROFILE = new Vector(-(float)Math.sqrt(2),(float)Math.sqrt(2));
+        public static final Vector LEFT_WHEEL_PAIR_PROFILE = new Vector(Math.sqrt(2),Math.sqrt(2));
+        public static final Vector RIGHT_WHEEL_PAIR_PROFILE = new Vector(-Math.sqrt(2),Math.sqrt(2));
     }
 
     public static final class AUTOMATED_CONTROLLER_PARAMS {
         public static final PIDFCalculator ROBOT_ROTATION_PID = new PIDFCalculator(1.1f, 0.0f, 100, 0.09f, 0, 0, 0, true);
         public static final PIDFCalculator ROBOT_PURSUIT_PID = new PIDFCalculator(0.0002f, 0f, 100, -0.00003f, 0, 0, 0, false);
-        public static final float STANDARD_PURSUIT_RANGE = 1300;
-        public static final float STANDARD_HEADING_RANGE = 0.08f;
+        public static final double STANDARD_PURSUIT_RANGE = 1300;
+        public static final double STANDARD_HEADING_RANGE = 0.08f;
         public static final int MAX_PURSUIT_SPEED = 10;
-        public static final float MAX_PURSUIT_ANGULAR_SPEED = 0.02f;
+        public static final double MAX_PURSUIT_ANGULAR_SPEED = 0.02f;
     }
 }
