+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/util/PIDFCalculator.java
@@ -1,20 +1,20 @@
 package org.firstinspires.ftc.teamcode.util;
 
 public class PIDFCalculator {
-    public float P;
-    public float I;
-    public float I_MAX;
-    public float D;
-    public float F;
-    public float FOffset;
-    public float FDividend;
+    public double P;
+    public double I;
+    public double I_MAX;
+    public double D;
+    public double F;
+    public double FOffset;
+    public double FDividend;
     public boolean constrainDifference;
 
-    private float lastState = 0;
-    public float target = 0;
-    private float error = 0;
+    private double lastState = 0;
+    public double target = 0;
+    private double error = 0;
 
-    public PIDFCalculator(float p, float i, float i_max, float d, float f, float fOffset, float fDividend, boolean _constrainDifference) {
+    public PIDFCalculator(double p, double i, double i_max, double d, double f, double fOffset, double fDividend, boolean _constrainDifference) {
         P = p;
         I = i;
         I_MAX = i_max;
@@ -25,16 +25,16 @@ public class PIDFCalculator {
         constrainDifference = _constrainDifference;
     }
 
-    public float evaluate(float currentState) {
-        float difference = constrainDifference ? DukUtilities.wrappedAngleDifference(target, currentState) : target - currentState;
-        float delta = constrainDifference ? DukUtilities.wrappedAngleDifference(lastState, currentState) : currentState - lastState;
-        float p = P * difference;
-        float i = I * error;
-        float d = D * delta / (float)TimeManager.getDeltaTime();
-        float f = F * -(float)Math.sin((currentState - FOffset) * (2 * Math.PI / FDividend));
+    public double evaluate(double currentState) {
+        double difference = constrainDifference ? DukUtilities.wrappedAngleDifference(target, currentState) : target - currentState;
+        double delta = constrainDifference ? DukUtilities.wrappedAngleDifference(lastState, currentState) : currentState - lastState;
+        double p = P * difference;
+        double i = I * error;
+        double d = D * delta / TimeManager.getDeltaTime();
+        double f = F * -Math.sin((currentState - FOffset) * (2 * Math.PI / FDividend));
 
         lastState = currentState;
-        error = (float)DukUtilities.clamp(error + difference, I_MAX, -I_MAX);
+        error = DukUtilities.clamp(error + difference, I_MAX, -I_MAX);
 
         return p + i + d + f;
     }
